# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: EMMA API
  version: 0.1.0
  description: OpenAPI specification for the EMMA API (developer-docs snapshot)
servers:
  - url: http://localhost:5262
    description: Local development
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    OrganizationCreateDto:
      type: object
      required: [name, email, ownerUserId]
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        ownerUserId:
          type: string
          format: uuid
        planId:
          type: string
          nullable: true
        planType:
          type: string
          nullable: true
        seatCount:
          type: integer
          minimum: 1
    OrganizationReadDto:
      type: object
      properties:
        id: { type: string, format: uuid }
        orgGuid: { type: string, format: uuid }
        name: { type: string }
        email: { type: string, format: email }
        planId: { type: string, nullable: true }
        planType: { type: string, nullable: true }
        seatCount: { type: integer, nullable: true }
        isActive: { type: boolean }
    InvitationCreateDto:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
        role:
          type: string
        expiresInDays:
          type: integer
        invitedByUserId:
          type: string
          format: uuid
    RegisterFromInvitationDto:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        fullName:
          type: string
        password:
          type: string
...
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
paths:
...
  /api/Organization:
    get:
      summary: List organizations (paged)
      tags: [Organization]
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema: { type: integer, minimum: 1, default: 1 }
        - in: query
          name: size
          schema: { type: integer, minimum: 1, default: 20, maximum: 100 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/OrganizationReadDto' }
              example:
                - id: "aaaaaaaa-aaaa-aaaa-aaaa-aaaaaaaaaaaa"
                  orgGuid: "bbbbbbbb-bbbb-bbbb-bbbb-bbbbbbbbbbbb"
                  name: "Acme Corp"
                  email: "owner@acme.com"
                  planId: "pro"
                  planType: "Pro"
                  seatCount: 10
                  isActive: true
        '401': { description: Unauthorized, content: { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } } } }
        '500': { description: Error, content: { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } } } }
    post:
      summary: Create organization
      tags: [Organization]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizationCreateDto'
            example:
              name: "Acme Corp"
              email: "owner@acme.com"
              ownerUserId: "22222222-2222-2222-2222-222222222222"
              planType: "Pro"
              seatCount: 10
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationReadDto'
              example:
                id: "aaaaaaaa-aaaa-aaaa-aaaa-aaaaaaaaaaaa"
                orgGuid: "bbbbbbbb-bbbb-bbbb-bbbb-bbbbbbbbbbbb"
                name: "Acme Corp"
                email: "owner@acme.com"
                planId: "pro"
                planType: "Pro"
                seatCount: 10
                isActive: true
        '400': { description: Bad Request, content: { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' }, example: { title: "Validation failed", status: 400, detail: "Name is required." } } } }
        '401': { description: Unauthorized, content: { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } } } }
        '409': { description: Conflict, content: { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' }, example: { title: "Conflict", status: 409, detail: "Organization with this name already exists." } } } }
        '500': { description: Error, content: { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } } } }
  /api/Organization/{id}:
    get:
      summary: Get organization by id
      tags: [Organization]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationReadDto'
        '401': { description: Unauthorized, content: { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } } } }
        '403': { description: Forbidden, content: { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } } } }
        '404': { description: Not Found, content: { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } } } }
  /api/Organization/{orgId}/invitations:
    post:
      summary: Create invitation for an organization
      tags: [Organization]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: orgId
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvitationCreateDto'
            example:
              email: "invitee@example.com"
              role: "Member"
              expiresInDays: 7
              invitedByUserId: "22222222-2222-2222-2222-222222222222"
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvitationCreateDto'
              example:
                id: "77777777-7777-7777-7777-777777777777"
                organizationId: "aaaaaaaa-aaaa-aaaa-aaaa-aaaaaaaaaaaa"
                email: "invitee@example.com"
                role: "Member"
                token: "4f3f0a2f5c0249e7a9c9e3f0a2f5c024"
                expiresAt: "2025-09-01T00:00:00Z"
                acceptedAt: null
                revokedAt: null
        '400': { description: Bad Request, content: { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' }, example: { title: "Validation failed", status: 400, detail: "Valid email is required" } } } }
        '401': { description: Unauthorized, content: { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } } } }
        '403': { description: Forbidden, content: { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' }, example: { title: "Forbidden", status: 403, detail: "Only owner or admin can invite." } } } }
        '409': { description: Conflict, content: { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' }, example: { title: "Conflict", status: 409, detail: "An active invitation already exists for this email" } } } }
        '500': { description: Error, content: { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } } } }
  /api/Organization/invitations/{token}:
    get:
      summary: Get invitation by token
      tags: [Organization]
      parameters:
        - in: path
          name: token
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
              example:
                id: "77777777-7777-7777-7777-777777777777"
                organizationId: "aaaaaaaa-aaaa-aaaa-aaaa-aaaaaaaaaaaa"
                email: "invitee@example.com"
                role: "Member"
                token: "4f3f0a2f5c0249e7a9c9e3f0a2f5c024"
                expiresAt: "2025-09-01T00:00:00Z"
                acceptedAt: null
                revokedAt: null
                isActive: true
        '400': { description: Bad Request, content: { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' }, example: { title: "Validation failed", status: 400, detail: "Token is required" } } } }
        '404': { description: Not Found, content: { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } } } }
        '500': { description: Error, content: { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } } } }
  /api/Organization/invitations/{token}/accept:
    post:
      summary: Accept invitation by token
      tags: [Organization]
      parameters:
        - in: path
          name: token
          required: true
          schema: { type: string }
      responses:
        '204': { description: No Content }
        '400': { description: Bad Request, content: { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' }, example: { title: "Validation failed", status: 400, detail: "Token is required" } } } }
        '404': { description: Not Found, content: { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' }, example: { title: "Not Found", status: 404, detail: "Invitation not found" } } } }
        '409': { description: Conflict, content: { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } } } }
        '500': { description: Error, content: { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } } } }
  /api/Organization/invitations/{token}/register:
    post:
      summary: Register from invitation token
      tags: [Organization]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterFromInvitationDto'
      parameters:
        - in: path
          name: token
          required: true
          schema: { type: string }
      responses:
        '204': { description: No Content }
        '400': { description: Bad Request, content: { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' }, example: { title: "Validation failed", status: 400, detail: "Token is required" } } } }
        '404': { description: Not Found, content: { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' }, example: { title: "Not Found", status: 404, detail: "Invitation not found" } } } }
        '409': { description: Conflict, content: { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } } } }
        '500': { description: Error, content: { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } } } }
...